requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
prometheus_client==0.20.0
PyYAML==6.0.1
cryptography==43.0.0
uwsgi==2.0.26